Enable `cargo run` on macOS with FSF toolchains
On macOS, the simulator has to be able to find `libgnat` at runtime -
which is provided by a macOS-specific environment variable that was not
automatically set. This update to `build.rs` ensures that `libgnat` can
be resolved automatically by `cargo run`.

With this change, a user can:
* clone the repo
* `cargo run --release` in `simulator`

and the demo runs successfully on macOS, with no further
configuration/setup required.