Improve the proofs for Rover.Autonomous
We now use a model of vehicle dynamics to prove `Go_Forward` in
`Rover.Autonomous` will not violate `Rover.Cannot_Crash`. We also
enhanced `Find_New_Direction` so that the Rover will confirm it is clear
to proceed along its new orientation, after choosing a new direction -
and loop until a clear direction is found.